add search box
