+++ libs/localization/src/lib/en.json
@@ -648,6 +648,7 @@
     "saveAsNew": "Save as new datapoint",
     "seePrediction": "See prediction deltas",
     "createOwn": "Create your own counterfactual:",
+    "noFeatures": "No features available",
     "referenceDatapoint": "Reference datapoint: Row {0}",
     "counterfactualEx": "Counterfactual Ex {0}",
     "saveDescription": "Without strict mathematical robustness of causal inference, we do not advise users to make real-life decisions based on this tool. Use our toolkit for causal analysis for real-life intervention.",
